+++ src/Components/Authentication/Login.js
@@ -1,7 +1,8 @@
 import React, { useRef, useState } from "react"
 import { Form, Button, Card, Alert } from "react-bootstrap"
-import { useAuth } from "../Contexts/AuthContext"
+import { useAuth } from "../../Contexts/AuthContext"
 import { Link, useHistory } from "react-router-dom"
+import CenteredContainer from './CenteredContainer'
 
 export default function Login() {
   const emailRef = useRef()
@@ -27,7 +28,7 @@ export default function Login() {
   }
 
   return (
-    <>
+    <CenteredContainer>
       <Card>
         <Card.Body>
           <h2 className="text-center mb-4">Log In</h2>
@@ -53,6 +54,6 @@ export default function Login() {
       <div className="w-100 text-center mt-2">
         Need an account? <Link to="/signup">Sign Up</Link>
       </div>
-    </>
+    </CenteredContainer>
   )
 }
