+++ src/Components/Authentication/UpdateProfile.js
@@ -1,7 +1,8 @@
 import React, { useRef, useState } from "react"
 import { Form, Button, Card, Alert } from "react-bootstrap"
-import { useAuth } from "../Contexts/AuthContext"
+import { useAuth } from "../../Contexts/AuthContext"
 import { Link, useHistory } from "react-router-dom"
+import CenteredContainer from './CenteredContainer'
 
 export default function UpdateProfile() {
   const emailRef = useRef()
@@ -43,7 +44,7 @@ export default function UpdateProfile() {
   }
 
   return (
-    <>
+    <CenteredContainer>
       <Card>
         <Card.Body>
           <h2 className="text-center mb-4">Update Profile</h2>
@@ -83,6 +84,6 @@ export default function UpdateProfile() {
       <div className="w-100 text-center mt-2">
         <Link to="/">Cancel</Link>
       </div>
-    </>
+    </CenteredContainer>
   )
 }
