+++ src/Components/Dnd/Profile.js
@@ -1,9 +1,9 @@
 import React, { useState } from "react"
 import { Card, Button, Alert } from "react-bootstrap"
-import { useAuth } from "../Contexts/AuthContext"
+import { useAuth } from "../../Contexts/AuthContext"
 import { Link, useHistory } from "react-router-dom"
 
-export default function Dashboard() {
+export default function Profile() {
   const [error, setError] = useState("")
   const { currentUser, logout } = useAuth()
   const history = useHistory()
